Gonna test the updated texture caching

+++ crates/defaults/src/components/sky.rs
@@ -3,7 +3,7 @@ use ecs::*;
 // A component that will be linked to the skysphere
 #[derive(Default)]
 pub struct Sky {
-    pub sky_gradient_texture_id: u16,
+    pub sky_gradient_texture_id: usize,
 }
 
 // Main traits implemented
